Add pull-to-refresh and scroll effects demos
Pull-to-refresh:  A simple pull-to-refresh demo with animations that are
in sync with scroll or triggered by it.

Scroll effects: A demo of multiple scroll synced effects composing
together.

There is also a simple wrapper that adds syntax sugar around current
scroll customization API to make it look like the ideal version we
eventually want.